Added per biome tree replacement rate config
Added per biome tree replacement rate config. Tried to respect code
style as much as i could.
Populating biomes map in init phase is important, as we are sure that in
init, all biomes have been registered. We are not for pre init.
Uses biomes registry and name mapping.
Default values not provided, but can be.